+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Breaking Changes
 
+-   **terra-tag** removed deprecated `inputBadge`. Use `name` instead.
 -   **terra-filter** removed deprecated `terra-filter`. Use the new `tc-filter` instead.
 -   **terra-slider** removed deprecated fields `inputValue` and `inputValueChange`.
 -   **terra-checkbox** removed getter and setter for `value` and `valueChange` emitter. Use `ngModel` and `ngModelChange` instead.